requirements.txt
django-environ-0.9.0
pip install Pillow